migrate to EUMM
At the cost of adding a prereq against Time::Piece everwhere
(even Unix where we don't actually need it) we are switching to
MakeMaker.  Since Time::Piece is in core as of Perl 5.10 (also
appears to be in 5.8.9) and of sufficient version in core at
that point it seems worth it.